+++ test/sql/linker.sql
@@ -55,3 +55,15 @@ select dlvalue(filename)
    and not (filename like '%X%' or filename like '%3.txt');
 
 truncate my_table2;
+
+-- test for foreign servers
+select * from datalink.curl_get('file://zala/etc/issue');
+create extension dblink;
+select * from datalink.curl_get('file://zala/etc/issue');
+create extension postgres_fdw;
+create server zala foreign data wrapper postgres_fdw;
+select * from datalink.curl_get('file://zala/etc/issue');
+select * from datalink.curl_get('file://tiha/etc/issue');
+create user mapping for current_role server zala;
+select url,ok,rc,error from datalink.curl_get('file://zala/etc/issue');
+select url,ok,rc,error from datalink.curl_get('file://zala/etc/issueXXXXX');